requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
Werkzeug==3.1.7
